+++ app/policies/ordering_configuration/service_context_policy.rb
@@ -2,6 +2,6 @@
 
 class OrderingConfiguration::ServiceContextPolicy < ServiceContextPolicy
   def show?
-    super && record.service.administered_by?(user)
+    super && record.service.administered_by?(user) && record.service.upstream&.eosc_registry?
   end
 end