+++ spec/features/service_spec.rb
@@ -12,11 +12,14 @@ RSpec.feature "Service browsing" do
   end
 
   context "with JS:" do
-    scenario "I can see 'Manage the resource' button if i am an admin provider", js: true do
+    scenario "I can see 'Manage the resource' button if i am an admin provider and it is an eosc_registry", js: true do
       user = create(:user)
       admin = create(:data_administrator, first_name: user.first_name, last_name: user.last_name, email: user.email)
       provider = create(:provider, data_administrators: [admin])
       service = create(:service, resource_organisation: provider)
+      service_source = create(:eosc_registry_service_source, service: service)
+      service.upstream = service_source
+      service.save!
       offer1 = create(:offer, service: service)
 
       checkin_sign_in_as(user)
@@ -94,10 +97,28 @@ RSpec.feature "Service browsing" do
       expect(page.body).not_to have_content("Service offers")
     end
 
-    scenario "I cannot see 'Manage the resource' button if i am not an admin provider" do
+    scenario "I cannot see 'Manage the resource' button if i am not an admin provider and it's an eosc_registry" do
       service = create(:service)
       offer1 = create(:offer, service: service)
+      service_source = create(:eosc_registry_service_source, service: service)
+      service.upstream = service_source
+      service.save!
 
+      visit service_path(service)
+
+      expect(page).to have_link("Access the resource")
+      expect(page).to_not have_content("Manage the resource")
+      expect(page).to_not have_content(offer1.name)
+    end
+
+    scenario "I cannot see 'Manage the resource' button if i am an admin provider but it's not eosc_registry" do
+      user = create(:user)
+      admin = create(:data_administrator, first_name: user.first_name, last_name: user.last_name, email: user.email)
+      provider = create(:provider, data_administrators: [admin])
+      service = create(:service, resource_organisation: provider)
+      offer1 = create(:offer, service: service)
+
+      checkin_sign_in_as(user)
       visit service_path(service)
 
       expect(page).to have_link("Access the resource")